Refactor TraitCatalog to explicitly instantiate SingleResourceCatalogs
We break out the definition of the SingleResourceTraitCatalog<> into a
separate header with an implementation in
SingleResourceTraitCatalog.ipp file, and explicitly template
instations in SingleResourceTraitCatalog.cpp.  The refactoring results
in a cleaner naming convention, faster compilation times and better
isolation of the SingleResourceTraitCatalog (instantiations are now
included in libWeave.a rather than in the applications using Weave)
